+++ plugins/vimeo.py
@@ -17,7 +17,7 @@ class VimeoConverter(BaseConverter):
     def get_capabilities(self) -> PluginCapabilities:
         return PluginCapabilities(
             name="Vimeo Converter",
-            version="1.5.0",
+            version="1.5.2",
             platform="Vimeo",
             description="Download videos from Vimeo",
             author="YouTube MP3 Downloader",